modularize new quote button
