+++ src/App.js
@@ -1,11 +1,12 @@
 import React, {useState, useContext} from 'react';
-import './App.css';
+import {Context} from './Context.js';
 import QuoteBox from './components/QuoteBox.js';
-import {ColorContext} from './ColorContext.js';
+import './App.css';
 import './RandomColor.scss'
 
 function App() {
-  const [colorNumber, setColorNumber] = useContext(ColorContext);
+  const {colorNum, quote, auth} = useContext(Context);
+  const [colorNumber, setColorNumber] = colorNum;  
   
   return (
     <div className={`App BG-color${colorNumber}`}>
